# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignPermissionsDto": {
        "additionalProperties": false,
        "properties": {
          "permissionIds": {
            "description": "Array of permission IDs to assign to the role",
            "example": [
              1,
              2,
              3
            ],
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "permissionIds"
        ],
        "type": "object"
      },
...
      "CheckPermissionDto": {
        "additionalProperties": false,
        "properties": {
          "discordId": {
            "description": "Discord member ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "permission": {
            "description": "Permission name/code to evaluate",
            "example": "ORGANIZER",
            "type": "string"
          },
          "serverId": {
            "description": "Server (guild) ID context",
            "example": "987654321098765432",
            "type": "string"
          }
        },
        "required": [
          "discordId",
          "serverId",
          "permission"
        ],
        "type": "object"
      },
      "CheckPermissionsBatchDto": {
        "additionalProperties": false,
        "properties": {
          "discordId": {
            "description": "Discord member ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "mode": {
            "description": "ALL = user must hold every permission, ANY = at least one",
            "enum": [
              "ALL",
              "ANY"
            ],
            "example": "ALL",
            "type": "string"
          },
          "permissions": {
            "description": "List of permission keys to check",
            "example": [
              "MANAGE_EVENTS",
              "SEND_MESSAGES"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "serverId": {
            "description": "Server (guild) ID context",
            "example": "987654321098765432",
            "type": "string"
          }
        },
        "required": [
          "discordId",
          "serverId",
          "permissions",
          "mode"
        ],
        "type": "object"
      },
...
      "ImpactPreviewDto": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "description": "Whether to preview adding or removing the permissions",
            "enum": [
              "add",
              "remove"
            ],
            "example": "add",
            "type": "string"
          },
          "permissionIds": {
            "description": "Array of permission IDs to preview impact for",
            "example": [
              1,
              2
            ],
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "permissionIds",
          "action"
        ],
        "type": "object"
      },
      "ImpactPreviewResponseDto": {
        "additionalProperties": false,
        "properties": {
          "affectedMembers": {
            "description": "Number of members affected by the change",
            "example": 15,
            "type": "number"
          },
          "memberIds": {
            "description": "Discord IDs of affected members",
            "example": [
              "111222333444555666",
              "777888999000111222"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "roleHolders": {
            "description": "Total number of members holding this role",
            "example": 15,
            "type": "number"
          }
        },
        "required": [
          "affectedMembers",
          "memberIds",
          "roleHolders"
        ],
        "type": "object"
      },
...
      "PermissionItemDto": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "description": "Permission description",
            "example": "Allows reading member data",
            "type": "string"
          },
          "id": {
            "description": "Permission ID",
            "example": 1,
            "type": "number"
          },
          "key": {
            "description": "Permission key",
            "example": "READ_MEMBERS",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key"
        ],
        "type": "object"
      },
...
      "RolePermissionsResponseDto": {
        "additionalProperties": false,
        "properties": {
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionItemDto"
            },
            "type": "array"
          },
          "roleId": {
            "description": "Discord role ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "roleName": {
            "description": "Role name",
            "example": "Developer",
            "type": "string"
          },
          "serverId": {
            "description": "Server ID",
            "example": "987654321098765432",
            "type": "string"
          }
        },
        "required": [
          "roleId",
          "roleName",
          "serverId",
          "permissions"
        ],
        "type": "object"
      },
...
      "UpsertInheritanceRuleDto": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "example": true,
            "type": "boolean"
          },
          "sourceRoleId": {
            "description": "Role ID from main server",
            "example": "112233445566778899",
            "type": "string"
          },
          "targetScope": {
            "enum": [
              "all",
              "selected"
            ],
            "example": "all",
            "type": "string"
          },
          "targetServerIds": {
            "description": "Required when targetScope = selected",
            "example": [
              "998877665544332211",
              "887766554433221100"
            ],
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "sourceRoleId",
          "targetScope"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "description": "Project API key (prefix.secret format)",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "session-token": {
        "bearerFormat": "Token",
        "description": "Enter your session token (from admin login or POST /auth/validate)",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "MicroClub Discord Identity API.\n\n**MC Project endpoints** authenticate with `X-API-Key` header.\n\n**Admin endpoints** authenticate with `Authorization: Bearer <token>`.\n\nScopes and operations are granted per project–server pair by a system administrator.",
    "title": "MCDI API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/permissions/admin/servers/{serverId}/roles/{roleId}/impact": {
      "post": {
        "description": "Read-only endpoint that returns how many members would be affected by adding or removing permissions from a role. Does not modify any data.",
        "operationId": "PermissionsController_previewImpact",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord role snowflake ID",
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "example": "987654321098765432",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpactPreviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactPreviewResponseDto"
                }
              }
            },
            "description": "Impact preview returned."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Preview impact of a permission change on a role",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/admin/servers/{serverId}/roles/{roleId}/permissions": {
      "get": {
        "description": "Returns the list of permissions currently assigned to a specific role in a server.",
        "operationId": "PermissionsController_getRolePermissions",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord role snowflake ID",
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "example": "987654321098765432",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionsResponseDto"
                }
              }
            },
            "description": "Role permissions list returned."
          },
          "400": {
            "description": "Invalid path parameter format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Get all permissions assigned to a role",
        "tags": [
          "Permissions"
        ]
      },
      "post": {
        "description": "Assigns one or more permissions to a role. Cache is invalidated immediately after the change.",
        "operationId": "PermissionsController_assignPermissionsToRole",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord role snowflake ID",
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "example": "987654321098765432",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionsResponseDto"
                }
              }
            },
            "description": "Updated role permissions list returned."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Add permissions to a role",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/admin/servers/{serverId}/roles/{roleId}/permissions/{permissionId}": {
      "delete": {
        "description": "Removes a single permission from a role. Executive roles (global or highest-ranking) are protected from modification. Cache is invalidated immediately after the change.",
        "operationId": "PermissionsController_removePermissionFromRole",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord role snowflake ID",
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "example": "987654321098765432",
              "type": "string"
            }
          },
          {
            "description": "Permission ID to remove",
            "in": "path",
            "name": "permissionId",
            "required": true,
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission removed successfully."
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid path parameter format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required, or role is protected from modification."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Remove a permission from a role",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/check": {
      "post": {
        "description": "Resolves whether a Discord member holds a specific permission in a given server. Permissions are derived from their roles and any configured inheritance rules. Requires the `check_permissions` scope for the server in the request body.",
        "operationId": "PermissionsController_checkPermission",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPermissionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission check result returned."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Valid API key required."
          },
          "403": {
            "description": "Server is disabled or project has no access."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Check a single permission",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/check-batch": {
      "post": {
        "description": "Checks a list of permissions for a member in one call. Set `mode` to `ALL` to require every permission, or `ANY` to pass if at least one matches. Requires the `check_permissions` scope for the server in the request body.",
        "operationId": "PermissionsController_checkPermissionBatch",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPermissionsBatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch permission check result returned."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Valid API key required."
          },
          "403": {
            "description": "Server is disabled or project has no access."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Check multiple permissions (batch)",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/inheritance-rules": {
      "get": {
        "description": "Returns all configured role inheritance rules.",
        "operationId": "PermissionsController_listInheritanceRules",
        "parameters": [
          {
            "description": "Filter by source role ID.",
            "in": "query",
            "name": "sourceRoleId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Filter by enabled status.",
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "description": "Filter by target scope type.",
            "in": "query",
            "name": "targetScope",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "selected"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter rules that apply to this server (scope=all always matches; scope=selected matches when this server is among targetServerIds).",
            "in": "query",
            "name": "serverId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inheritance rules list returned."
          },
          "400": {
            "description": "Invalid parameter format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "List inheritance rules",
        "tags": [
          "Permissions"
        ]
      },
      "post": {
        "description": "Upserts a permission inheritance rule that causes members holding a source role to also receive all permissions of a target role, optionally scoped to a specific server.",
        "operationId": "PermissionsController_upsertInheritanceRule",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertInheritanceRuleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Inheritance rule upsert result returned."
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Create or update an inheritance rule",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/{serverId}/{discordId}": {
      "get": {
        "description": "Returns every effective permission the member holds in the server, including permissions inherited via role hierarchy rules. Requires the `check_permissions` scope for this server.",
        "operationId": "PermissionsController_getMemberPermissions",
        "parameters": [
          {
            "description": "Discord guild snowflake ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord user snowflake ID",
            "in": "path",
            "name": "discordId",
            "required": true,
            "schema": {
              "example": "876543210987654321",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full permission set returned."
          },
          "400": {
            "description": "Invalid path parameter format."
          },
          "401": {
            "description": "Valid API key required."
          },
          "403": {
            "description": "Server is disabled or project has no access."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get full resolved permissions for a member",
        "tags": [
          "Permissions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API Server",
      "url": "http://localhost:3000"
    }
  ],
...
}